// CaLOLateClient/CaLOLateClient/src/app/interfaces/Player.ts
import { Player } from './PlayerElement';

export class PlayerClass implements Player{
    name: string = "";
    level: Number = 0;
    healthPer: Number = 0;
    item1: string = "";
    item2: string = "";
    item3: string = "";
    item4: string = "";
    item5: string = "";
    item6: string = "";
    constructor(){}
  
  }